Tensorflow 2 Unet implement done

+++ Unet/unet.py
@@ -13,10 +13,13 @@ class Unet():
 
     def create(self):
         self.model = tf.keras.Sequential()
-        input = tf.keras.layers.Input(shape=(28, 28, 1))
+        input = tf.keras.layers.Input(shape=(28, 28, 1), dtype=tf.float32)
         x = input
         x, skip_layers = self.addNpass_down_path(x)
         x = self.add_up_path(x, skip_layers)
+
+        # residual learning
+        # x = x + input
 
         return tf.keras.Model(inputs=input, outputs=x)
